bot/requirements.txt
# Core
python-binance>=1.0.19
uvloop>=0.21.0; sys_platform != "win32"

# API Framework
fastapi>=0.115.0
uvicorn[standard]>=0.34.0

# Data & Indicators (no pandas_ta - implemented with pure numpy/pandas)
pandas>=2.2.0
numpy>=1.26.0

# Models & Validation
pydantic>=2.10.0
pydantic-settings>=2.7.0

# Database
aiosqlite>=0.20.0

# AI
anthropic>=0.39.0

# HTTP
httpx>=0.28.0
aiohttp>=3.11.0

# Utilities
python-dotenv>=1.0.1
orjson>=3.10.0
